requirements.txt
jinja2
pyyaml
rjsmin
csscompressor
